pyproject.toml
[project]
name = "lx200emulator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"

requires-python = ">=3.14"
dependencies = [
    "pyserial>=3.5",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
markers = [
    "unit: unit tests (parser, policy, reducer); no TCP or server.",
]
addopts = "-v"
